chores: modularised the function to extract scores from html, so it could be used elsewhere
